@@ -9,63 +9,74 @@
     "css": "sass"
   },
   "scripts": {
-    "dev:weapp": "pnpm run build:weapp --watch",
-    "uat:weapp": "cross-env MODE=uat pnpm run build:weapp",
-    "build:weapp": "taro build --type weapp",
+    "dev": "pnpm run build --watch",
+    "build": "taro build --type weapp",
+    "develop": "pnpm run build --mode develop",
     "dev:h5": "pnpm run build:h5 --watch",
-    "uat:h5": "cross-env MODE=uat pnpm run build:h5",
     "build:h5": "taro build --type h5",
     "preview:h5": "serve -s dist",
     "type:check": "tsc --noEmit",
     "lint": "eslint --fix ."
   },
   "dependencies": {
-    "@nutui/nutui-taro": "3.1.20",
-    "@tarojs/components": "3.4.8",
-    "@tarojs/taro": "3.4.8",
+    "@abraham/reflection": "^0.10.0",
+    "@babel/runtime": "^7.18.9",
+    "@nutui/nutui-taro": "^3.1.8",
+    "@tarojs/components": "3.5.4",
+    "@tarojs/helper": "3.5.4",
+    "@tarojs/plugin-framework-vue3": "3.5.4",
+    "@tarojs/plugin-html": "3.5.4",
+    "@tarojs/plugin-platform-weapp": "3.5.4",
+    "@tarojs/router": "3.5.4",
+    "@tarojs/runtime": "3.5.4",
+    "@tarojs/shared": "3.5.4",
+    "@tarojs/taro": "3.5.4",
+    "@tarojs/taro-h5": "3.5.4",
     "injection-js": "^2.4.0",
+    "path-to-regexp": "^6.2.1",
     "taro-axios": "^1.1.1",
-    "ts-essentials": "^9.1.2",
+    "ts-essentials": "^9.3.0",
     "tslib": "^2.4.0",
-    "vue": "^3.2.33",
-    "vue3-oop": "^0.5.8"
+    "vue": "^3.2.37",
+    "vue3-oop": "^1.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/runtime": "^7.17.9",
-    "@commitlint/cli": "^16.2.4",
-    "@commitlint/config-conventional": "^16.2.4",
+    "@babel/core": "^7.18.13",
+    "@commitlint/cli": "^17.1.1",
+    "@commitlint/config-conventional": "^17.1.0",
     "@nexhome/yorkie": "^2.0.8",
-    "@tarojs/cli": "3.4.8",
-    "@tarojs/mini-runner": "3.4.8",
-    "@tarojs/plugin-framework-vue3": "^3.4.8",
-    "@tarojs/plugin-html": "3.4.8",
-    "@tarojs/runner-utils": "^3.4.8",
-    "@tarojs/runtime": "3.4.8",
-    "@tarojs/webpack-runner": "3.4.8",
-    "@types/webpack-env": "^1.16.4",
-    "@typescript-eslint/eslint-plugin": "^5.22.0",
-    "@typescript-eslint/parser": "^5.22.0",
-    "@vue3-oop/jsx-hot-loader": "^0.1.6",
-    "@vue3-oop/taro-plugin": "^0.1.0",
-    "autoprefixer": "^10.4.7",
-    "babel-plugin-import": "^1.13.5",
-    "babel-preset-taro": "3.4.8",
-    "commitizen": "^4.2.4",
+    "@tarojs/cli": "3.5.4",
+    "@tarojs/plugin-mock": "^0.0.9",
+    "@tarojs/webpack5-runner": "3.5.4",
+    "@types/node": "^18.7.13",
+    "@types/webpack-env": "^1.18.0",
+    "@typescript-eslint/eslint-plugin": "^5.35.1",
+    "@typescript-eslint/parser": "^5.35.1",
+    "@vue/babel-plugin-jsx": "^1.0.6",
+    "@vue/compiler-sfc": "^3.2.37",
+    "@vue3-oop/taro-plugin": "^1.0.2",
+    "autoprefixer": "^10.4.8",
+    "babel-plugin-import": "^1.13.3",
+    "babel-preset-taro": "3.5.4",
     "cross-env": "^7.0.3",
-    "eslint": "^8.14.0",
+    "css-loader": "6.7.1",
+    "eslint": "~8.22.0",
     "eslint-config-prettier": "^8.5.0",
+    "eslint-config-taro": "3.5.4",
     "eslint-plugin-prettier": "^4.0.0",
-    "lint-staged": "^12.4.1",
+    "eslint-plugin-vue": "^9.4.0",
+    "lint-staged": "^13.0.3",
+    "postcss": "^8.4.16",
     "prettier": "^2.6.2",
-    "serve": "^13.0.2",
-    "taro-plugin-tailwind": "^1.2.8",
-    "ts-deepmerge": "^2.0.1",
+    "serve": "^14.0.1",
+    "style-loader": "3.3.1",
+    "stylelint": "14.11.0",
     "ts-gear": "^4.7.3",
-    "typescript": "^4.6.4",
-    "vite": "^2.9.8",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.2",
+    "vite": "^3.0.9",
     "vue-loader": "^17.0.0",
-    "windicss": "^3.5.1"
+    "webpack": "5.74.0"
   },
   "commitlint": {
     "extends": [
