@@ -1,4 +1,4 @@
-import 'windi.css'
+import '@abraham/reflection'
 import './theme/app.scss'
 import { createApp } from 'vue'
 
