@@ -1,23 +1,13 @@
-import defaultConf from './config.default'
-import developmentConf from './config.development'
-import productionConf from './config.production'
-import uatConf from './config.uat'
-import deepMerge from 'ts-deepmerge'
+import Config from '@/config/config.default'
 
-let conf = defaultConf
-const mergeOpt = { mergeArrays: false }
+let TargetConf = Config
+// 自动扫描配置文件
+const requireContext = require.context('./', false, /config\.(\w+)\.ts$/)
+requireContext.keys().forEach(path => {
+  if (path !== `./config.${process.env.VUE_APP_MODE}.ts`) return
+  TargetConf = requireContext(path).default
+})
 
-switch (process.env.MODE) {
-  case 'development':
-    conf = deepMerge.withOptions(mergeOpt, defaultConf, developmentConf)
-    break
-  case 'uat':
-    conf = deepMerge.withOptions(mergeOpt, defaultConf, uatConf)
-    break
-  case 'production':
-    conf = deepMerge.withOptions(mergeOpt, defaultConf, productionConf)
-    break
-}
-conf.env = process.env.MODE
+const config = new TargetConf()
 
-export default conf
+export default config
